+++ src/api/schema.ts
@@ -95,6 +95,11 @@ export interface CanvasContext {
  */
 export interface MediaContext {
   /**
+   * 0-based index of the current active media file in the list of media files in the media
+   * response.
+   */
+  index?: number
+  /**
    * Media progress of current active media file.
    */
   progress?: string
@@ -975,6 +980,11 @@ export interface ListItem {
  */
 export interface Media {
   /**
+   * 0-based index of the first Media Object in media_objects to play. If unspecified, zero or
+   * out-of-bounds, playback starts at the first media object in media_objects.
+   */
+  firstMediaObjectIndex?: number
+  /**
    * List of Media Objects
    */
   mediaObjects?: MediaObject[]
@@ -985,6 +995,10 @@ export interface Media {
    * two default control type, FINISHED and FAILED.
    */
   optionalMediaControls?: OptionalMediaControl[]
+  /**
+   * Repeat mode for the list of Media Objects.
+   */
+  repeatMode?: RepeatMode
   /**
    * Start offset of the first media object.
    */
@@ -1037,6 +1051,15 @@ export enum OptionalMediaControl {
   OptionalMediaControlsUnspecified = 'OPTIONAL_MEDIA_CONTROLS_UNSPECIFIED',
   Paused = 'PAUSED',
   Stopped = 'STOPPED',
+}
+
+/**
+ * Repeat mode for the list of Media Objects.
+ */
+export enum RepeatMode {
+  All = 'ALL',
+  Off = 'OFF',
+  RepeatModeUnspecified = 'REPEAT_MODE_UNSPECIFIED',
 }
 
 /**